requirements.txt
streamlit==1.12.0
    mysql-connector-python==8.0.31
    requests==2.28.1
    translate==3.6.0
